requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-multipart==0.0.9
playwright==1.47.0
pypdf==4.3.1
pymupdf>=1.24.10
paddlepaddle==2.6.2
paddleocr==2.8.1
easyocr==1.7.1
pdf2image==1.17.0
surya-ocr==0.6.3
rapidfuzz==3.9.7
langdetect==1.0.9
pillow==10.4.0
numpy==1.26.4
shapely==2.0.6
supabase==2.7.4
python-dotenv==1.0.1
pydantic==2.8.2
loguru==0.7.2
